requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.31.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.65.0 